requirements.txt
beautifulsoup4==4.11.1
matplotlib>=3.7
nltk>=3.7
openai==1.97.1
pandas==1.5.3
prettytable==3.16.0
PyYAML==6.0
scikit_learn==1.2.1
spacy==3.7.4
tqdm>=4.64
aiofiles==24.1.0
chardet==4.0.0
charset_normalizer==3.3.2
docx2txt==0.8
matplotlib==3.8.3
networkx==3.2.1
numpy==1.23.5
playwright==1.54.0
pygraphviz==1.13
python_Levenshtein==0.27.1
PyYAML==6.0
Requests==2.32.4
scipy==1.10.0
seaborn==0.13.2
thefuzz==0.22.1
tiktoken==0.7.0
